+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project follows Semantic Versioning.
 
 ---
 
+## [2.1.2] - 2026-04-12
+
+### Improvements
+- Added light/dark mode toggle
+- Changed 'wet track' to a slider
+
+
 ## [2.1.1] - 2026-03-30
 
 ### Fixes